Reverted back to 3bits per vram/buckbuffer number (so you can select 0 to 7 each, although no more than total 6 are possible to fit in VRAM now) as there was still an effect I was thinking to do that will need 2 main VRAM buffers (for double buffering) and 4 additional backbuffer (to store as many previous frames as possible).
Meanwhile, I was trying to call CreateBasicDisplay and if it returns an error to retry calling it after decreasing the number of total buffers. That didn't work for some reason. But now I made a function that only checks arithmetically if the total buffers num by the screen size in bytes exceeds the available VRAM memory. If so, we try again with less backbuffers and also less vram buffers. I can ask for 7 VRAM buffers and 7 backbuffers and end up automatically with 6 and 0. Works like a charm!